Encapsulate AI metrics parsing inside each AiPlugin (#242)
Move per-tool session-metrics wire-format parsing out of session_metrics.rs
and into plugins/ai/{claude,copilot,codex}/metrics.rs behind a generic
MetricsParser trait. session_metrics.rs keeps only the generic engine
(run_metrics_watcher): threading, polling, file tailing with truncation/
rotation handling, discovery backoff, snapshot dedup, and emit plumbing.

- Add MetricsParser trait + LocatedFile + un_metrics_watcher, replacing
  run_claude/copilot/codex_watcher and the MetricsWatcherKind enum dispatch.
- AiPlugin::metrics_watcher_kind -> metrics_parser(...) -> Option<Box<dyn MetricsParser>>.
- Strategy flags relocate_each_poll (Claude) and rebind_on_disappear (Codex)
  preserve each tool's exact original behavior; no change to surfaced metrics.
- Extract shared context_used_pct() helper (was triplicated, overflow-safe u128).
- Move per-tool tests alongside their parsers; add codex engine integration test.
